+++ Makefile
@@ -12,6 +12,7 @@ FILES  = ./configParsing/getLocation.cpp \
 		./Response/requese.cpp \
 		./Response/response.cpp \
 		./Response/tools.cpp \
+		./Response/tools2.cpp \
 		./serveurs-clients/serverTools.cpp \
 		./serveurs-clients/servers2.cpp
