+++ Response/response.cpp
@@ -15,10 +15,14 @@ std::string Response::build_response()
         this->other_response("505", "Version Not Supported");
     else if (this->status == 405)
         this->other_response("405", "Method not allowed");
+    else if (this->status == 414)
+        this->other_response("405", "URI Too Long");
     else if (this->status == 413)
         this->other_response("413", "Request-URI Too Long");
     else if (this->status == 411)
         this->other_response("411", "Length Required");
+    else if (this->status == 501)
+        this->other_response("501", "Not Implemented");
     else if (!this->response_items.location->return_code_url.empty())
     {
         std::string url = (this->response_items.location->root + this->response_items.Path);
@@ -34,46 +38,7 @@ std::string Response::build_response()
     return (response.str());
 }
 
-std::string Response::get_Content_type(std::string url)
-{
 
-    std::string contentTypes[25] = {
-        "text/plain",
-        "text/html",
-        "text/css",
-        "text/javascript",
-        "text/xml",
-        "application/json",
-        "application/xml",
-        "application/xhtml+xml",
-        "application/pdf",
-        "application/octet-stream",
-        "image/jpeg",
-        "image/png",
-        "image/jpg",
-        "image/gif",
-        "image/svg+xml",
-        "audio/mpeg",
-        "audio/wav",
-        "video/mp4",
-        "video/webm",
-        "multipart/form-data",
-        "multipart/byteranges",
-        "application/x-www-form-urlencoded",
-        "application/graphql",
-        "application/vnd.api+json",
-    };
-    std::string extension = url.substr(url.rfind(".") + 1);
-    unsigned int i = 0;
-    while (i < 25)
-    {
-        if (contentTypes[i].find(extension) != std::string::npos)
-            return contentTypes[i];
-        i++;
-    }
-
-    return "text/html;";
-}
 
 void Response::build_GET()
 {
@@ -164,16 +129,12 @@ void Response::build_GET()
                             else if (this->get_permission(URI) == -2)
                                 this->other_response("403", " Forbidden");
                             else
-                            {
-                                // std::string content_body = read_file(URI);
                                 this->ft_success_code("200",  read_file(URI), URI);
-                            }
                         }
                     }
                     else
                     {
                         URI += "index.html";
-                        // std::string content_body = read_file(URI.c_str());
                         if (access(URI.c_str(), F_OK | W_OK) != 0)
                         {
                             this->other_response("403", " Forbidden");
@@ -195,10 +156,7 @@ void Response::build_GET()
                 else if (this->get_permission(URI) == -2)
                     this->other_response("403", " Forbidden");
                 else
-                {
-                    // std::string content_body = read_file(URI);
                     this->ft_success_code("200", read_file(URI), URI);
-                }
             }
         }
     }
@@ -260,23 +218,28 @@ void Response::build_DELETE()
 void Response::build_POST()
 {
 
-       
     struct stat buffer;
     int status;
-    std::string upload_enable = this->response_items.location->upload_enable; // replace it by location value
+    std::string upload_enable = this->response_items.location->upload_enable;
     std::string upload_store_directory = this->response_items.location->upload_store_directory;
-    std::string URI = this->response_items.location->root;                 // change by value depends on location
-    std::string get_auto_index = this->response_items.location->autoindex; // change by getter
-    std::string cgi_path = this->response_items.location->cgi_path;        // change path with valid path from config;
-
-    URI += this->response_items.Path.substr(1);
-
+    std::string URI = this->response_items.location->root;                 
+    std::string get_auto_index = this->response_items.location->autoindex; 
+    std::string cgi_path = this->response_items.location->cgi_path;        
     std::string index;
     int pos = 0;
     unsigned int k = 0;
     std::stringstream ss;
+    const size_t BUFFER_SIZE = 3000;
+    std::string namefile;
+    time_t current_time;
+    std::ofstream file;
+    std::vector<RequestBody *>::iterator it;
+
+
+    URI += this->response_items.Path.substr(1);
 
     status = stat(URI.c_str(), &buffer);
+    std::cout << URI.c_str() << std::endl;
     if (upload_enable == "off")
     {
         if (status != -1)
@@ -286,12 +249,15 @@ void Response::build_POST()
                 if (!cgi_path.empty())
                     responsecgi(GET_CGI_DATA(this->response_items));
                 else
+                {
+
                     this->other_response("403", " Forbidden");
+                }
             }
             else
             {
                 if (this->response_items.Path[this->response_items.Path.size() - 1] == '/')
-                    this->ft_redirect("300", this->response_items.Path + "/");
+                    this->ft_redirect("300", this->response_items.Path);
                 else
                 {
                     index = this->check_index_file(URI);
@@ -317,10 +283,7 @@ void Response::build_POST()
     }
     else
     {
-        std::string namefile;
-        time_t current_time;
-        std::ofstream file;
-        std::vector<RequestBody *>::iterator it;
+        
         if (this->response_items.Headers.find("Content-Type")->second.find("multipart/form-data") != std::string::npos)
         {
             it = this->response_items.ChunkedBody.begin();
@@ -353,7 +316,6 @@ void Response::build_POST()
                     }
                     if (!(*it)->Content.empty())
                     {
-                        const size_t BUFFER_SIZE = 3000;
                         size_t totalBytesWritten = 0;
                         while (totalBytesWritten < (*it)->Content.size()) {
                             size_t bytesToWrite = std::min(BUFFER_SIZE, (*it)->Content.size() - totalBytesWritten);
@@ -393,36 +355,9 @@ void Response::build_POST()
     }
 }
 
-int Response::get_permission(std::string &file)
-{
-    if (access(file.c_str(), F_OK) == -1)
-        return -1;
-    if (access(file.c_str(), R_OK) == -1)
-        return -2;
-    return 0;
-}
-
-std::string Response::read_file(const std::string &filename)
-{
-    std::ifstream file(filename);
-    std::stringstream ss;
-    if (!file.is_open())
-    {
-        return "not found";
-    }
-    ss << file.rdbuf();
-    file.close();
-    return ss.str();
-}
-
-
-
-
-
 
 void Response::return_pages(std::string &pages_return, std::string &url)
 {
-    // std::cout << "Returning" << std::endl;
     std::vector<std::string> pages = split_v(pages_return, " ");
     switch (atoi(pages[0].c_str()))
     {
@@ -450,23 +385,11 @@ void Response::return_pages(std::string &pages_return, std::string &url)
     }
 }
 
-void Response::ft_default_pages(std::string status, std::string &message, std::string &path)
-{
-    struct stat buffer;
-    std::string filename;
-    if (stat(this->response_items.error_pages.find(status)->second.c_str(), &buffer) != -1)
-    {
-        // if (path[0] != '/')
-        //     path = "/" + path;
-        
-        filename = path;
-        message = this->read_file(filename);
-    }
-}
+
 
 void Response::ft_success_code(std::string status, std::string message, std::string URI)
 {
-    std::string connection = (!this->response_items.Headers["Connection:"].empty() ? this->response_items.Headers["Connection:"] : "close");
+    std::string connection = (!this->response_items.Headers["Connection"].empty() ? this->response_items.Headers["Connection"] : "close");
     if (this->response_items.error_pages.find(status) != this->response_items.error_pages.end())
         ft_default_pages(status, message, (this->response_items.error_pages.find(status)->second));
     response << "HTTP/1.1 " << status << " ok\r\n";
@@ -481,7 +404,7 @@ void Response::ft_success_code(std::string status, std::string message, std::str
 
 void Response::ft_redirect(std::string status, std::string message)
 {
-     std::string connection = (!this->response_items.Headers["Connection:"].empty() ? this->response_items.Headers["Connection:"] : "close");
+     std::string connection = (!this->response_items.Headers["Connection"].empty() ? this->response_items.Headers["Connection"] : "close");
     if (this->response_items.error_pages.find(status) != this->response_items.error_pages.end())
     {
         ft_default_pages(status, message, (this->response_items.error_pages.find(status)->second));
@@ -499,7 +422,7 @@ void Response::ft_redirect(std::string status, std::string message)
 void Response::other_response(std::string status, std::string desc)
 {
     std::string body;
-    std::string connection = (!this->response_items.Headers["Connection:"].empty() ? this->response_items.Headers["Connection:"] : "close");
+    std::string connection = (!this->response_items.Headers["Connection"].empty() ? this->response_items.Headers["Connection"] : "close");
     body = "<!DOCTYPE html>\n<html lang=\"en\">\n\
                                                     <head>\n\
                                                     <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n";
