+++ file.conf
@@ -1,6 +1,6 @@
 server  {
   listen 8006;
- #root /Users/ichouare/Desktop/webserv42/src;
+ root /Users/ichouare/Desktop/webserv42/src;
   upload_store_directory /Users/ichouare/Desktop/webserv42/fortest;                                      
   # listening port, mandatory parameter fgfg
   host 127.0.0.1;                    
@@ -61,7 +61,7 @@ server  {
                     #   ftjj                # EX: - URI /tours           --> docs/fusion_web/tours               
                                       #     - URI /tours/page.html --> docs/fusion_web/tours/page.html 
       autoindex off;
-      upload_enable on;
+      upload_enable off;
 
       # return 200 "hello world!";              
       # turn on/off directory listing
@@ -101,6 +101,7 @@ server  {
       allow POST  GET DELETE;
       autoindex on;
       cgi_path  /usr/bin/php;
+      upload_enable off;
       # location of interpreters installed on the current system, mandatory parameter
       cgi_extension .php;                               
       # extensions for executable files, mandatory parameter
