+++ src/index.php
@@ -1,0 +1,12 @@
+<?php
+$handle = fopen("cgi_in", "r");
+if ($handle) {
+    while (($line = fgets($handle)) !== false) {
+        echo $line;
+    }
+
+    fclose($handle);
+} else {
+    die("Unable to open stdin!");
+}
+?>
