--- src/websites/index.php
@@ -1,3 +1,0 @@
-<?php
-print "Hello, World!";
-?>
